+++ src/main/scala/nl/knaw/dans/easy/stage/dataset/AdditionalLicense.scala
@@ -1,7 +1,8 @@
-package nl.knaw.dans.easy.stage
+package nl.knaw.dans.easy.stage.dataset
 
 import java.io.File
 
+import nl.knaw.dans.easy.stage.Settings
 import nl.knaw.dans.easy.stage.lib.Constants
 import org.apache.commons.io.FileUtils
